gtk overrides: Add support for builder scope
This is a follow of the previous commit which fixes template support
in GTK4 by adding GtkBuilerScope. It is extended to support a scope
object (it can be a python object or a dictionnary). This scope object
is used to define the scope the builder should operate in. A new
GtkBuilder constructor is introduced to be able to define this scope
object.

The different relevant tests are updated to support both GTK3 and
GTK4.